+++ tools/data/kinetics/download_annotations.sh
@@ -1,6 +1,13 @@
 #!/usr/bin/env bash
 
 DATASET=$1
+if [ "$DATASET" == "kinetics400" ] || [ "$1" == "kinetics600" ] || [ "$1" == "kinetics700" ]; then
+        echo "We are processing $DATASET"
+else
+        echo "Bad Argument, we only support kinetics400, kinetics600 or kinetics700"
+        exit 0
+fi
+
 DATA_DIR="../../../data/${DATASET}/annotations"
 
 if [[ ! -d "${DATA_DIR}" ]]; then