+++ tools/data/kinetics/download_videos.sh
@@ -6,6 +6,12 @@ source activate kinetics
 pip install --upgrade youtube-dl
 
 DATASET=$1
+if [ "$DATASET" == "kinetics400" ] || [ "$1" == "kinetics600" ] || [ "$1" == "kinetics700" ]; then
+        echo "We are processing $DATASET"
+else
+        echo "Bad Argument, we only support kinetics400, kinetics600 or kinetics700"
+        exit 0
+fi
 
 DATA_DIR="../../../data/${DATASET}"
 ANNO_DIR="../../../data/${DATASET}/annotations"